NCBC-2176: SDK 3 Migration Guide
Motivation
----------
Migrating to SDK 3 is desirable for more features, less surface area
and the better users are able to do this, the less we need to maintain
old versions

Modifications
-------------
Starting with the Java SDK as a guide, authored many migration guide
sections. Did not complete all sections, but stopped with enough to
be relevant.

Results
-------
Users can read and migrate, or they will find it with a search.