DOC-9636: Document new RawBinaryTranscoder features in .NET SDK 3.2.6
Motivation
----------
As of .NET SDK 3.2.6 the RawBinaryTranscoder supports additional types
besides byte[] that offer greater performance when dealing with large
binary objects.

Modifications
-------------
Add documentation and an example that cover encoding of `Memory<byte>`
and `ReadOnlyMemory<byte>` and decoding of `IMemoryOwner<byte>`.

Results
-------
This "hidden" performance feature is now documented.